+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## Unreleased
 
 - Added `zip` option
+- Added support to store `Infinity`
 - Dropped support for Node.js below version 8